Add code for ClientException, correct attribute inheritance and move it to internal.
