+++ Neo4j.Driver/Neo4j.Driver/Exceptions/ClientException.cs
@@ -1,14 +1,14 @@
 ﻿// Copyright (c) "Neo4j"
 // Neo4j Sweden AB [http://neo4j.com]
-// 
+//
 // This file is part of Neo4j.
-// 
+//
 // Licensed under the Apache License, Version 2.0 (the "License"):
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 //     http://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,6 +17,7 @@
 
 using System;
 using System.Runtime.Serialization;
+using Neo4j.Driver.Internal.ExceptionHandling;
 
 namespace Neo4j.Driver;
 
@@ -25,6 +26,7 @@ namespace Neo4j.Driver;
 /// The error code provided can be used to determine further detail for the problem.
 /// </summary>
 [DataContract]
+[ClientErrorCode("Neo.ClientError.*")]
 public class ClientException : Neo4jException
 {
     /// <summary>
