+++ Neo4j.Driver/Neo4j.Driver/Exceptions/ForbiddenException.cs
@@ -16,6 +16,7 @@
 // limitations under the License.
 
 using System.Runtime.Serialization;
+using Neo4j.Driver.Internal.ExceptionHandling;
 
 namespace Neo4j.Driver;
