+++ Neo4j.Driver/Neo4j.Driver/Exceptions/TokenExpiredException.cs
@@ -15,6 +15,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Neo4j.Driver.Internal.ExceptionHandling;
+
 namespace Neo4j.Driver;
 
 /// <summary>
